# docs/openapi.yaml
openapi: 3.1.0
info:
  title: DAM API
  version: 1.0.0
  description: |
    Digital Asset Management API. All authenticated routes require either a
    JWT Bearer token (obtained via `/auth/login`) or an API key passed as
    `X-API-Key: <key>` (or `?api_key=<key>` query param).

    Multi-tenancy is enforced server-side — every entity is scoped to the
    authenticated org. The org ID is derived from the JWT/API-key claims or
    the custom `Host:` header for CDN requests.
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - BearerAuth: []
  - ApiKeyAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    Org:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        plan:
          type: string
        storage_quota_bytes:
          type: integer
          format: int64
        storage_used_bytes:
          type: integer
          format: int64
        settings:
          $ref: '#/components/schemas/OrgSettings'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        org_id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/Role'
        active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    TokenPair:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string

    # ── Common list wrappers ──────────────────────────────────────────────────
...
paths:
...
  /api/v1/auth/register:
    post:
      tags: [auth]
      summary: Register a new org and owner account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [org_name, email, password]
              properties:
                org_name:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/TokenPair'
                  - type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
                      org:
                        $ref: '#/components/schemas/Org'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/auth/login:
    post:
      tags: [auth]
      summary: Log in and receive JWT tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/TokenPair'
                  - type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/auth/refresh:
    post:
      tags: [auth]
      summary: Exchange a refresh token for a new token pair
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: New token pair
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/auth/logout:
    post:
      tags: [auth]
      summary: Revoke the current refresh token
      responses:
        '204':
          description: Logged out
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ── Assets ─────────────────────────────────────────────────────────────────
  /api/v1/assets:
    get:
      tags: [assets]
      summary: List assets (cursor-paginated)
      parameters:
        - name: q
          in: query
          schema:
            type: string
          description: Full-text search query
        - name: folder_id
          in: query
          schema:
            type: string
            format: uuid
        - name: mime_group
          in: query
          schema:
            type: string
            enum: [image, video, document]
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 200
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [created_at, name, size, updated_at]
        - name: sort_dir
          in: query
          schema:
            type: string
            enum: [asc, desc]
      responses:
        '200':
          description: Paginated asset list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [assets]
      summary: Upload a single asset
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                folder_id:
                  type: string
                  format: uuid
                visibility:
                  $ref: '#/components/schemas/Visibility'
                title:
                  type: string
                description:
                  type: string
                alt_text:
                  type: string
                author:
                  type: string
      responses:
        '201':
          description: Uploaded asset
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '402':
          description: Storage quota exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/assets/bulk:
    post:
      tags: [assets]
      summary: Upload multiple assets in one request
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [files]
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '201':
          description: Bulk upload results
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Asset'
                  errors:
                    type: array
                    items:
                      type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/assets/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [assets]
      summary: Get an asset by ID
      responses:
        '200':
          description: Asset
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [assets]
      summary: Update asset metadata and/or visibility
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                metadata:
                  $ref: '#/components/schemas/AssetMetadata'
                visibility:
                  $ref: '#/components/schemas/Visibility'
      responses:
        '200':
          description: Updated asset
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [assets]
      summary: Soft-delete an asset
      responses:
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/assets/{id}/move:
    post:
      tags: [assets]
      summary: Move an asset to a different folder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                folder_id:
                  type: string
                  format: uuid
                  nullable: true
                  description: Target folder ID, or null for root
      responses:
        '204':
          description: Moved
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/assets/{id}/signed-url:
    get:
      tags: [assets]
      summary: Generate a time-limited signed URL for a private asset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: ttl
          in: query
          schema:
            type: integer
            default: 3600
          description: TTL in seconds
      responses:
        '200':
          description: Signed URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                    format: uri
                  expires_in:
                    type: number
                    description: TTL in seconds
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/assets/{id}/tags:
    post:
      tags: [assets]
      summary: Add a tag to an asset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tag_id]
              properties:
                tag_id:
                  type: string
                  format: uuid
      responses:
        '204':
          description: Tag added
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [assets]
      summary: Remove a tag from an asset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tag_id]
              properties:
                tag_id:
                  type: string
                  format: uuid
      responses:
        '204':
          description: Tag removed
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Folders ────────────────────────────────────────────────────────────────
  /api/v1/folders:
    get:
      tags: [folders]
      summary: List folders
      parameters:
        - name: parent_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Folder list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Folder'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [folders]
      summary: Create a folder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                parent_id:
                  type: string
                  format: uuid
                  nullable: true
      responses:
        '201':
          description: Created folder
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Folder'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/folders/tree:
    get:
      tags: [folders]
      summary: Get the full folder tree
      responses:
        '200':
          description: Nested folder tree
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Folder'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/folders/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [folders]
      summary: Get a folder by ID
      responses:
        '200':
          description: Folder
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Folder'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [folders]
      summary: Rename a folder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        '200':
          description: Updated folder
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Folder'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [folders]
      summary: Delete a folder
      responses:
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Tags ───────────────────────────────────────────────────────────────────
  /api/v1/tags:
    get:
      tags: [tags]
      summary: List tags for the org
      responses:
        '200':
          description: Tag list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Tag'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [tags]
      summary: Create a tag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        '201':
          description: Created tag
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/tags/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    delete:
      tags: [tags]
      summary: Delete a tag
      responses:
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Collections ────────────────────────────────────────────────────────────
  /api/v1/collections:
    get:
      tags: [collections]
      summary: List collections
      responses:
        '200':
          description: Collection list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Collection'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [collections]
      summary: Create a collection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        '201':
          description: Created collection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Collection'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/collections/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [collections]
      summary: Get a collection by ID
      responses:
        '200':
          description: Collection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Collection'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [collections]
      summary: Update a collection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        '200':
          description: Updated collection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Collection'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [collections]
      summary: Delete a collection
      responses:
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/collections/{id}/assets:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [collections]
      summary: List assets in a collection
      responses:
        '200':
          description: Asset list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/collections/{id}/assets/{assetId}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: assetId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    post:
      tags: [collections]
      summary: Add an asset to a collection
      responses:
        '204':
          description: Added
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [collections]
      summary: Remove an asset from a collection
      responses:
        '204':
          description: Removed
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Image Styles ───────────────────────────────────────────────────────────
  /api/v1/styles:
    get:
      tags: [styles]
      summary: List image styles
      responses:
        '200':
          description: Style list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ImageStyle'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [styles]
      summary: Create an image style (owner/admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, slug]
              properties:
                name:
                  type: string
                slug:
                  type: string
                operations:
                  type: array
                  items:
                    $ref: '#/components/schemas/StyleOperation'
                output_format:
                  $ref: '#/components/schemas/OutputFormat'
                quality:
                  type: integer
                  minimum: 1
                  maximum: 100
      responses:
        '201':
          description: Created style
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageStyle'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /api/v1/orgs/me:
    get:
      tags: [orgs]
      summary: Get the current org
      responses:
        '200':
          description: Org
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Org'
        '401':
          $ref: '#/components/responses/Unauthorized'

    put:
      tags: [orgs]
      summary: Update org settings (owner/admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                settings:
                  $ref: '#/components/schemas/OrgSettings'
      responses:
        '200':
          description: Updated org
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Org'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...